Implement kernel-level exception handler.
This necessitated changes to the wait and stop semantics, as documented
in the `Kernel` interface API documentation.
