+++ test/suite/React/ReactKernel.spec.php
@@ -10,7 +10,6 @@ use React\EventLoop\Factory;
 use React\EventLoop\LoopInterface;
 use Recoil\Kernel\Api;
 use Recoil\Recoil;
-use RuntimeException;
 use Throwable;
 
 describe(ReactKernel::class, function () {
@@ -43,17 +42,6 @@ describe(ReactKernel::class, function () {
             })->to->throw(
                 Exception::class,
                 '<exception>'
-            );
-        });
-
-        it('detects abandoned coroutines', function () {
-            expect(function () {
-                ReactKernel::start(function () {
-                    yield Recoil::suspend();
-                });
-            })->to->throw(
-                RuntimeException::class,
-                'The strand never exited.'
             );
         });
 
@@ -95,39 +83,6 @@ describe(ReactKernel::class, function () {
         it('runs the event loop', function () {
             $this->subject->wait();
             $this->eventLoop->run->called();
-        });
-
-        it('can be invoked again after an interrupt', function () {
-            $this->eventLoop->run->does(function () {
-                $this->subject->interrupt(new Exception());
-            });
-
-            expect(function () {
-                $this->subject->wait();
-            })->to->throw(Exception::class);
-
-            expect(function () {
-                $this->subject->wait();
-            })->to->be->ok;
-        });
-    });
-
-    describe('->interrupt()', function () {
-        it('stops the event loop', function () {
-            $exception = Phony::mock(Throwable::class)->mock();
-            $this->subject->interrupt($exception);
-            $this->eventLoop->stop->called();
-        });
-
-        it('causes wait() to throw', function () {
-            $exception = Phony::mock(Throwable::class)->mock();
-            $this->eventLoop->run->does(function () use ($exception) {
-                $this->subject->interrupt($exception);
-            });
-
-            expect(function () {
-                $this->subject->wait();
-            })->to->throw($exception);
         });
     });
